+++ stories/02-Design System/03-Navigation/SideMenu/examples/index.tsx
@@ -4,6 +4,7 @@ import menuDisabled from './SideMenu/menuDisabled';
 import menuSearch from './SideMenu/menuSearch';
 import menuSeparator from './SideMenu/menuSeparator';
 import menuComplement from './SideMenu/menuComplement';
+import menuCustomContent from './SideMenu/menuCustomContent';
 
 import menuBasicSource from './SideMenu/menuBasic?raw';
 import menuActiveSource from './SideMenu/menuActive?raw';
@@ -11,6 +12,7 @@ import menuDisabledSource from './SideMenu/menuDisabled?raw';
 import menuSearchSource from './SideMenu/menuSearch?raw';
 import menuSeparatorSource from './SideMenu/menuSeparator?raw';
 import menuComplementSource from './SideMenu/menuComplement?raw';
+import menuCustomContentSource from './SideMenu/menuCustomContent?raw';
 
 export const Sources = {
     menuBasic: menuBasicSource,
@@ -18,7 +20,8 @@ export const Sources = {
     menuDisabled: menuDisabledSource,
     menuSearch: menuSearchSource,
     menuSeparator: menuSeparatorSource,
-    menuComplement: menuComplementSource
+    menuComplement: menuComplementSource,
+    menuCustomContent: menuCustomContentSource
 };
 
 export default {
@@ -27,5 +30,6 @@ export default {
     menuDisabled,
     menuSearch,
     menuSeparator,
-    menuComplement
+    menuComplement,
+    menuCustomContent
 };